Refactor - Split cardano-wallet.js into byron-tx-planner.js and max-amount-calculator.js
