+++ app/frontend/wallet/cardano-wallet.ts
@@ -1,8 +1,4 @@
-import {encode} from 'borc'
-import {base58} from 'cardano-crypto.js'
-
 import debugLog from '../helpers/debugLog'
-import {generateMnemonic, validateMnemonic} from './mnemonic'
 
 import {TxInputFromUtxo, TxOutput, TxAux} from './byron/byron-transaction'
 
@@ -10,19 +6,19 @@ import AddressManager from './address-manager'
 import {ByronAddressProvider} from './byron/byron-address-provider'
 import BlockchainExplorer from './blockchain-explorer'
 import PseudoRandom from './helpers/PseudoRandom'
-import {ADA_DONATION_ADDRESS, MAX_INT32, TX_WITNESS_SIZE_BYTES} from './constants'
+import {MAX_INT32} from './constants'
 import shuffleArray from './helpers/shuffleArray'
-import CborIndefiniteLengthArray from './byron/helpers/CborIndefiniteLengthArray'
 import NamedError from '../helpers/NamedError'
-import {roundWholeAdas} from '../helpers/adaConverters'
 import {Lovelace} from '../state'
 
-function txFeeFunction(txSizeInBytes: number): Lovelace {
-  const a = 155381
-  const b = 43.946
+import {computeRequiredTxFee, selectMinimalTxPlan, isUtxoProfitable} from './byron-tx-planner'
 
-  return Math.ceil(a + txSizeInBytes * b) as Lovelace
-}
+import {MaxAmountCalculator} from './max-amount-calculator'
+
+const {
+  getMaxDonationAmount: _getMaxDonationAmount,
+  getMaxSendableAmount: _getMaxSendableAmount,
+} = MaxAmountCalculator(computeRequiredTxFee)
 
 type UTxO = {
   txHash: string
@@ -31,58 +27,14 @@ type UTxO = {
   outputIndex: number
 }
 
-type Input = UTxO
+export type Input = UTxO
 
-type Output = {
+export type Output = {
   address: string
   coins: Lovelace
 }
 
-// Estimates size of final transaction in bytes.
-// Note(ppershing): can overshoot a bit
-function estimateTxSize(inputs: Array<Input>, outputs: Array<Output>): Lovelace {
-  // exact size for inputs
-  const preparedInputs = inputs.map(TxInputFromUtxo)
-  const txInputsSize = encode(new CborIndefiniteLengthArray(preparedInputs)).length
-
-  const maxCborCoinsLen = 9 //length of CBOR encoded 64 bit integer, currently max supported
-
-  const txOutputsSizes = outputs.map(
-    // Note(ppershing): we are conservative here
-    // FIXME(ppershing): shouldn't there be some +1 for the array encoding?
-    // Is it in maxCborCoinsLen?
-    ({address, coins}) => base58.decode(address).length + maxCborCoinsLen
-  )
-
-  // +2 for indef array start & end
-  const txOutputsSize = txOutputsSizes.reduce((acc, x) => acc + x, 0) + 2
-
-  const txMetaSize = 1 // currently empty Map
-
-  // the 1 is there for the CBOR "tag" for an array of 4 elements
-  const txAuxSize = 1 + txInputsSize + txOutputsSize + txMetaSize
-
-  const txWitnessesSize = inputs.length * TX_WITNESS_SIZE_BYTES + 1
-
-  // the 1 is there for the CBOR "tag" for an array of 2 elements
-  const txSizeInBytes = 1 + txAuxSize + txWitnessesSize
-
-  /*
-  * the slack is there for the array of tx witnesses
-  * because it may have more than 1 byte of overhead
-  * if more than 16 elements are present
-  */
-  const slack = 4
-
-  return txSizeInBytes + slack
-}
-
-function computeRequiredTxFee(inputs: Array<Input>, outputs: Array<Output>): Lovelace {
-  const fee = txFeeFunction(estimateTxSize(inputs, outputs))
-  return fee
-}
-
-interface TxPlan {
+export interface TxPlan {
   inputs: Array<Input>
   outputs: Array<Output>
   change: Output | null
@@ -91,63 +43,6 @@ interface TxPlan {
 
 interface NoTxPlan {
   estimatedFee: Lovelace
-}
-
-function computeTxPlan(
-  inputs: Array<Input>,
-  outputs: Array<Output>,
-  possibleChange: Output
-): TxPlan | null {
-  const totalInput = inputs.reduce((acc, input) => acc + input.coins, 0)
-  const totalOutput = outputs.reduce((acc, output) => acc + output.coins, 0)
-
-  if (totalOutput > Number.MAX_SAFE_INTEGER) {
-    throw NamedError('CoinAmountError')
-  }
-
-  const feeWithoutChange = computeRequiredTxFee(inputs, outputs)
-
-  // Cannot construct transaction plan
-  if (totalOutput + feeWithoutChange > totalInput) return null
-
-  // No change necessary, perfect fit
-  if (totalOutput + feeWithoutChange === totalInput) {
-    return {inputs, outputs, change: null, fee: feeWithoutChange as Lovelace}
-  }
-
-  const feeWithChange = computeRequiredTxFee(inputs, [...outputs, possibleChange])
-
-  if (totalOutput + feeWithChange > totalInput) {
-    // We cannot fit the change output into the transaction
-    // Instead, just increase the fee
-    return {
-      inputs,
-      outputs,
-      change: null,
-      fee: (totalOutput - totalInput) as Lovelace,
-    }
-  }
-
-  return {
-    inputs,
-    outputs,
-    change: {
-      address: possibleChange.address,
-      coins: (totalInput - totalOutput - feeWithChange) as Lovelace,
-    },
-    fee: feeWithChange as Lovelace,
-  }
-}
-
-function getUtxoBalance(utxos: Array<UTxO>): Lovelace {
-  return utxos.reduce((acc, utxo) => acc + utxo.coins, 0) as Lovelace
-}
-
-function isUtxoProfitable(utxo: UTxO) {
-  const inputSize = encode(TxInputFromUtxo(utxo)).length
-  const addedCost = txFeeFunction(inputSize + TX_WITNESS_SIZE_BYTES) - txFeeFunction(0)
-
-  return utxo.coins > addedCost
 }
 
 function prepareTxAux(plan: TxPlan) {
@@ -177,80 +72,6 @@ function filterUnusedEndAddresses(
     }
   }
   return addressesWithMeta.slice(0, minCount)
-}
-
-function _getMaxSendableAmount(utxos, address, hasDonation, donationAmount, donationType) {
-  const profitableUtxos = utxos.filter(isUtxoProfitable)
-  const coins = getUtxoBalance(profitableUtxos)
-
-  if (!hasDonation) {
-    const inputs = profitableUtxos
-    const outputs = [{address, coins: 0 as Lovelace}]
-
-    const txFee = computeRequiredTxFee(inputs, outputs)
-    return {sendAmount: Math.max(coins - txFee, 0)}
-  } else {
-    const inputs = profitableUtxos
-    const outputs = [
-      {address, coins: 0 as Lovelace},
-      {address: ADA_DONATION_ADDRESS, coins: 0 as Lovelace},
-    ]
-    const txFee = computeRequiredTxFee(inputs, outputs)
-
-    if (donationType === 'percentage') {
-      // set maxSendAmount and percentageDonation (0.2% of max) to deplete balance completely
-      const percent = 0.2
-
-      const reducedAmount: Lovelace = Math.floor(coins / (1 + percent / 100)) as Lovelace
-      const roundedDonation = roundWholeAdas(((reducedAmount * percent) / 100) as Lovelace)
-
-      return {
-        sendAmount: coins - txFee - roundedDonation,
-        donationAmount: roundedDonation,
-      }
-    } else {
-      return {sendAmount: Math.max(coins - donationAmount - txFee, 0)}
-    }
-  }
-}
-
-function _getMaxDonationAmount(utxos, address, sendAmount: Lovelace) {
-  const profitableUtxos = utxos.filter(isUtxoProfitable)
-  const coins = getUtxoBalance(profitableUtxos)
-
-  const inputs = profitableUtxos
-  const outputs = [
-    {address, coins: 0 as Lovelace},
-    {address: ADA_DONATION_ADDRESS, coins: 0 as Lovelace},
-  ]
-
-  const txFee = computeRequiredTxFee(inputs, outputs)
-  return Math.max(coins - txFee - sendAmount, 0)
-}
-
-function selectMinimalTxPlan(
-  utxos: Array<UTxO>,
-  address,
-  coins,
-  donationAmount,
-  changeAddress
-): TxPlan | NoTxPlan {
-  const profitableUtxos = utxos.filter(isUtxoProfitable)
-
-  const inputs = []
-
-  const outputs = [{address, coins}]
-  if (donationAmount > 0) outputs.push({address: ADA_DONATION_ADDRESS, coins: donationAmount})
-
-  const change = {address: changeAddress, coins: 0 as Lovelace}
-
-  for (let i = 0; i < profitableUtxos.length; i++) {
-    inputs.push(profitableUtxos[i])
-    const plan = computeTxPlan(inputs, outputs, change)
-    if (plan) return plan
-  }
-
-  return {estimatedFee: computeRequiredTxFee(inputs, outputs)}
 }
 
 const CardanoWallet = (options) => {
@@ -435,4 +256,4 @@ if (typeof window !== 'undefined') {
   window.CardanoWallet = CardanoWallet
 }
 
-export {CardanoWallet, generateMnemonic, validateMnemonic, txFeeFunction}
+export {CardanoWallet}
